tests/wpt/web-platform-tests/bluetooth/requestDevice/blocklisted-service-in-filter.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'Reject with SecurityError if requesting a blocklisted ' +
    'service.';
const expected = new DOMException(
    'requestDevice() called with a filter containing a blocklisted UUID. ' +
        'https://goo.gl/4NeimX',
    'SecurityError');

bluetooth_test(
    () =>
        setUpPreconnectedDevice({knownServiceUUIDs: ['human_interface_device']})
            .then(
                () => assert_promise_rejects_with_message(
                    requestDeviceWithTrustedClick(
                        {filters: [{services: ['human_interface_device']}]}),
                    expected, 'Requesting blocklisted service rejects.')),
    test_desc);
</script>
